sseserver/weather/pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.11.3",
    "langchain-mcp-adapters>=0.1.9",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.5",
    "python-dotenv>=1.1.1",
    "mcp[cli]>=1.13.0",
]
